+++ src/imports/utils/configutils.ts
@@ -1,18 +1,31 @@
 import * as vscode from 'vscode';
 
-export const config = vscode.workspace.getConfiguration('MythicScribe');
 
 export function isAlwaysEnabled() {
-    return config.get('alwaysEnabled');
+    return vscode.workspace.getConfiguration('MythicScribe').get('alwaysEnabled');
 }
 
 export function isEnabled(document: vscode.TextDocument): boolean {
     if (isAlwaysEnabled()) {
         return true;
     }
-    const path = document.uri.fsPath.toLowerCase();
-    if (path.includes('mythicmobs')) {
+    const regex: string|undefined = vscode.workspace.getConfiguration('MythicScribe').get<string>('regexForMythicmobsFile');
+    const path = document.uri.fsPath;
+    if (regex && new RegExp(regex).test(path)) {
         return true;
     }
     return false;
+}
+
+export function isMetaskillFile(document: vscode.TextDocument): boolean {
+    const regex: string|undefined = vscode.workspace.getConfiguration('MythicScribe').get<string>('regexForMetaskillFile');
+    const path = document.uri.fsPath;
+    if (regex && new RegExp(regex).test(path)) {
+        return true;
+    }
+    return false;
+}
+
+export function enableEmptyBracketsAutomaticRemoval() {
+    return vscode.workspace.getConfiguration('MythicScribe').get('enableEmptyBracketsAutomaticRemoval');
 }